✨ 为 mongodb 实现 `GetDatabaseConnect` trait
