+++ database/mongo_connection/src/lib.rs
@@ -1,8 +1,9 @@
-#![feature(generic_associated_types)]
 #![feature(type_alias_impl_trait)]
 mod config;
 mod database;
 mod error;
+mod impl_get_connect;
+mod impl_initial;
 mod mongo_connect;
 
 mod static_vars;
@@ -12,56 +13,9 @@ pub type MongoClient = mongodb::Client;
 pub type MongoErr = mongodb::error::Error;
 pub type MongoClientOptions = mongodb::options::ClientOptions;
 
-use std::future::Future;
-
 pub use config::{DbConnectConfig, MongoDbConfig};
 pub use database::manager::{CollectionGuard, DatabaseManage};
-use database_traits::initial::{
-    DatabaseInitial, DatabaseInitialBasic, DatabaseInitialConnect,
-    DatabaseInitialMigration,
-};
 pub use error::MongoDbError;
+pub use impl_get_connect::MongoConnect;
 pub use mongo_connect::MongoConnectBuilder;
 pub use static_vars::{get_mongo_collection, get_mongo_database};
-
-impl DatabaseInitialBasic for DatabaseManage {
-    type Builder = MongoConnectBuilder;
-    type Error = MongoErr;
-}
-
-impl<C> DatabaseInitialConnect<C> for DatabaseManage
-where
-    C: DbConnectConfig + 'static + Sized,
-{
-    type ConnectFuture<'p> =
-        impl Future<Output = Result<Self::Builder, Self::Error>> + 'p;
-
-    fn start_connect(params: &C) -> Self::ConnectFuture<'_> {
-        async { MongoConnectBuilder::new(params).await }
-    }
-}
-
-impl<'p, M> DatabaseInitialMigration<'p, M> for DatabaseManage
-where
-    M: mongo_migrate_util::MigratorTrait + Sync + Send + 'p,
-{
-    type MigrateFuture =
-        impl Future<Output = Result<Self::Builder, Self::Error>>;
-
-    fn apply_migration(
-        builder: Self::Builder, params: M,
-    ) -> Self::MigrateFuture {
-        async move { builder.apply_mongo_migration(params).await }
-    }
-}
-
-impl DatabaseInitial for DatabaseManage {
-    type BuildResult = ();
-
-    fn build(
-        builder: Self::Builder,
-    ) -> Result<Self::BuildResult, Self::Error> {
-        builder.build();
-        Ok(())
-    }
-}
